readability and create moduleRepository base class #77
