+++ src/feelpp/benchmarking/report/atomicReports/view.py
@@ -2,3 +2,4 @@ from feelpp.benchmarking.report.base.view import View
 
 class AtomicReportView(View):
     pass
+
